+++ src/parser.rs
@@ -10,7 +10,7 @@ use std::error::Error;
 use std::fmt::{self, Formatter, Write};
 use std::str;
 
-use form_urlencoded::query_encoding::{self, EncodingOverride};
+use form_urlencoded::EncodingOverride;
 use host::{Host, HostInternal};
 use percent_encoding::{percent_encode, utf8_percent_encode, AsciiSet, CONTROLS};
 use Url;
@@ -1441,7 +1441,11 @@ impl<'a> Parser<'a> {
             "http" | "https" | "file" | "ftp" | "gopher" => self.query_encoding_override,
             _ => None,
         };
-        let query_bytes = query_encoding::encode(encoding, &query);
+        let query_bytes = if let Some(o) = encoding {
+            o(&query)
+        } else {
+            query.as_bytes().into()
+        };
         let set = if scheme_type.is_special() {
             SPECIAL_QUERY
         } else {
